requirements.txt
telethon==1.41.2
python-dotenv==1.0.1
pydantic==2.9.2
aiosqlite==0.20.0
httpx==0.27.2
